Move core server types into types folder
