+++ src/common/server/index.ts
@@ -1,2 +1,2 @@
 export { DNSServer, DefaultServer } from './DefaultServer';
-export { DNSRequest, DNSResponse, Handler, NextFunction } from './types';
+export { DNSRequest, DNSResponse, Handler, NextFunction } from '../../types/server';
